restructure classes into client and server
