+++ src/edu/ucsb/cs56/games/client_server/v2/client/Controllers/TicTacToeController.java
@@ -3,7 +3,7 @@ package edu.ucsb.cs56.games.client_server.v2.client.Controllers;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaClient;
+//import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaClient;
 import edu.ucsb.cs56.games.client_server.v2.client.Views.TicTacToeViewPanel;
 
 /**
