+++ src/edu/ucsb/cs56/games/client_server/v2/client/Controllers/TwoPlayerGameController.java
@@ -1,6 +1,6 @@
 package edu.ucsb.cs56.games.client_server.v2.client.Controllers;
 
-import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaClient;
+//import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaClient;
 
 /**
  * Two Player Game Controller which should be extended from for all two player games.
