+++ src/edu/ucsb/cs56/games/client_server/v2/server/Controllers/ChatController.java
@@ -8,7 +8,7 @@ package edu.ucsb.cs56.games.client_server.v2.server.Controllers;
 
 import java.util.ArrayList;
 
-import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaServer;
+//import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaServer;
 
 /**
  * Chat service is a service that most services will extend, it provides functionality for handling messages and chat-related data
