+++ src/edu/ucsb/cs56/games/client_server/v2/server/Controllers/ServiceController.java
@@ -2,7 +2,7 @@ package edu.ucsb.cs56.games.client_server.v2.server.Controllers;
 
 import java.util.ArrayList;
 
-import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaServer;
+//import edu.ucsb.cs56.games.client_server.v2.Controllers.JavaServer;
 
 /**
  * Service is an abstract class to be extended by all services
